New backups (17+2r): Restoring data from the login screen (#3003)
* New backups: Restoring data from the login screen

This is work in progress. For time being I commented out encryption.

* bringing back kotlinx-coroutines-core
* TODO comments and excluding old coroutines library version

+++ common-test/src/main/kotlin/com/waz/zclient/framework/data/messages/LikesTestDataProvider.kt
@@ -6,7 +6,7 @@ import java.util.UUID
 data class LikesTestData(
     val messageId: String,
     val userId: String,
-    val timeStamp: Int,
+    val timeStamp: Long,
     val action: Int
 )
 
@@ -15,7 +15,7 @@ object LikesTestDataProvider : TestDataProvider<LikesTestData>() {
         LikesTestData(
             messageId = UUID.randomUUID().toString(),
             userId = UUID.randomUUID().toString(),
-            timeStamp = System.currentTimeMillis().toInt(),
+            timeStamp = System.currentTimeMillis(),
             action = 0
         )
 }
